livekit/pyproject.toml
[project]
name = "avaluma-livekit-examples"
version = "1.0.0"
description = "Simple examples for LiveKit Agents with Avaluma Avatar"
requires-python = ">=3.12"

dependencies = [
    "livekit",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
]
